fixed_dich

+++ render_utils.c
@@ -6,7 +6,7 @@
 /*   By: tyamcha <tyamcha@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/18 16:16:50 by tyamcha           #+#    #+#             */
-/*   Updated: 2021/12/18 19:13:14 by tyamcha          ###   ########.fr       */
+/*   Updated: 2021/12/23 14:04:20 by tyamcha          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,8 +44,8 @@ void	place(t_vars *vars, int i, int j)
 
 int	render_frame(t_vars *vars)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	while (i < vars->map_height)
